requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy==1.26.4
gdown
PyYAML
opencv-python-headless